vp9-svc: Fix to buffer update under frame_drops
For svc with frame dropping in full_superframe_drop or
constrained dropped mode: the buffer level for a given layer
may be capped from increasing too much because that layer may be
dropped even though its buffer is stable (the dropped is force
due to underflow in other layers).

The capping already exists, but we should also cap
the bits_off_target as well, since that is used for update on
next frame, which in turn affects qp for next frame/layer.

Change-Id: Ifdab5d478e91cce20ecec51faa574eed375ee36b