Update patch set 2
Patch Set 2:

Hi Jingning,

We found that kf_boost calculation is calculated over a maximum of 32 frames. This number could have been arrived as 2 times max_gf_interval. However, we feel that for simpler contents, boost should be calculated over larger window as coding key frame with better quality is helpful for more frames in the relevant key frame interval. Hence, we tried adapting the interval based on zero_motion_percent or predictability. 

Please find the BDRATE impact for cpu-used=1 below : (Negative value signifies improvement)

BDRATE (based on GLBPsnr) on an average is -0.23% ranging between 3.16% and -6.42%. 
BDRATE (based on AVGPsnr) on an average is -0.21% ranging between 3.35% and -4.85%.
BDRATE (based on VPXSSIM) on an average is -0.35% ranging between 3.90% and -8.35%.

The testing is performed for the suite shared on Jan 14th. Also we found that encode time reduction is negligible for touchdown_pass content at 5 mbps. However, when tested for FourPeople with 160, 320, 640, 1280, 2560 kpbs encode time reduction is seen as ~5% with more encode time reduction at low bit-rates.

When we submitted the patch, we found that the jenkins build has failed. We could not find the reason for the same and also we could not re-trigger the runs. Can you help us in finding the relevant issue.

Regards,
Deepa

Patch-set: 2
Reviewer: Gerrit User 1001914 <1001914@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1000884 <1000884@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1177966 <1177966@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1273230 <1273230@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>