Fix Windows SDK and VS version checks
If WindowsTargetPlatformVersion is not set, the Visual Studio 15 (2017)
toolchain assumes that Windows 8.1 is being targeted. Since ARM64
support is only present and unlocked in Windows SDKs >= Windows 10 1809,
set that SDK as required in the vcxproj files.

Note that Visual Studio is aware of this bug and planning a fix in
Visual Studio 16:

https://developercommunity.visualstudio.com/content/problem/128836/windowstargetplatformversion-to-use-the-latest-ava.html

Change-Id: Ib069501ad384d91349b1f635722dedd31a4edd97