Update patch set 1
Patch Set 1:

In thinking about it now this slightly violates the break between building the tools with their code against the public api only and not relying on internal headers. This can make it easier for people copying the example code. Maybe it's best to leave this one as is for now.

Patch-set: 1
Reviewer: Gerrit User 1002219 <1002219@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>