Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Just curious how much it helps the compression performance of AQ 1? Do you have borg test results?
> 
> It is more that the AQ1 was simply not working right since this code was added in. If you coded a clip like Akiyo most of the blocks ended up in segment 0. Essentially the bug was preventing the segments being allocated sensibly.

In VBR if all (or almost all) blocks are forced falsely to one segment then it will behave in a manner that is closer to non AQ mode so metrics may be better. Basically the rate control adjusts for the Q offset and it was "as if" AQ was off.

With this change it tests different partition sizes using the same (and hence fair) rd multiplier, rather than preferring a partition structure that pushes as many blocks as possible into the bottom segment due to it using a lower RD value.

So to take an example. If a 32x32 block has one 8x8 with low variance but the rest is high variance. The old code will prefer a 32x32 prediction block size over spiting down as this 32x32 block then gets assigned (according to the rules used) to the segment of the lowest sub block (the 8x8 block with low variance). Because this low variance segment has a lower rd multiplier it comes out looking better in the rd calculation than would be the case if the block was partitioned down into 1 8x8 in the low variance segment and the rest in the high variance segment (which is the outcome we actually want).

Patch-set: 2
Reviewer: Gerrit User 1000899 <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>