Set up Wiener variance for macroblocks in a frame
This commit introduces the Wiener variance term. For each block in
the source frame, we first estimate its film grain noise level
using median filter in the transform domain. Each transform
coefficient is then processed using Wiener filter to account for
the impact on the energy level due to film grain noise. The result
leads to a second moment of the true signal.

Change-Id: Ibce7cb1b0cb8fe1aba807d95289712271d576948