Update patch set 2
Patch Set 2:

Following is the BD-Rate impact of this patch-set (negative values signifies improvement, tested for cpu-used=1)

BD-Rate (based on FASTSIM) on an average is -1.04% ranging between 1.8% and -19.01%.

BD-Rate (based on VPXSSIM) on an average is -0.69% ranging between 1.90% and -11.96%.

BD-Rate (based on GLBPsnr) on an average is -0.48% ranging between 2.31% and -4.39%.

BD-Rate (based on AVGPsnr) on an average is -0.27% ranging between 2.76% and -2.75%.

Overall encode time reduction of 16.78% is seen for 1080p touchdown_pass content at 5 mbps. The reason seems to be because of better bit consumption in an ARF interval.

Patch-set: 2
Reviewer: Gerrit User 1328009 <1328009@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>