Rework tile boundary condition for intra prediction
Explicitly compare the block location against tile coordinate to
decide if intra prediction boundary is available. This fixes an
encoder issue that uses cross tile boundary data for intra prediction
in the rate-distortion optimizations.

For 4K Disney and Fox test clips coded in 8 tiles, it improves
the speed 0 performance in several chunks by 0.2 - 0.3 dB. On average
2.2% gains in PSNR.

Change-Id: I80b3a131366bb2c5f8ea53a139ed6e9b0b7ddb68