Update patch set 3
Patch Set 3:

Following is the BD-Rate impact of this patch-set (negative values signifies improvement, tested for cpu-used=1)

BD-Rate (based on VPXSSIM) on an average is -0.2% ranging between 1.6% and -3.6%.

BD-Rate (based on GLBPsnr) on an average is -0.13% ranging between 1.15% and -2.71%.

BD-Rate (based on AVGPsnr) on an average is -0.08% ranging between 1.13% and -2.71%.

Overall encode time increase of 8% is seen for 1080p touchdown_pass content at 5 mbps. Averaging the encode time for 3, 5, 8, 10 mbps, we observed no impact on  encoding time.

Patch-set: 3
Reviewer: Gerrit User 1328009 <1328009@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>