Fix timestamp overflow issues
- Save the initial user-specified timestamp and rebase all further
timestamps by this value. This makes libaom internal timestamps to
always start from zero, regardless of the user's timestamps.
- Calculate reduced timestamp conversion ratio and use it to convert
user's timestamps to libaom internal timestamps and back. The effect
of this is that integer overflow due to multiplication doesn't
happen for a much longer time.
- For TimestampTest, increase the starting timestamp so that the bug
can be reproduced before this patch, and enable the test.

BUG=webm:701

Change-Id: Ic6f5eacd9a7c21b95707d31ee2da77dc8ac7dccf