vp9-rtc: tx_size selection for intra mode in nonrd
In nonrd_pickmode for intra modes: refactor code to
add tx_size selection based on prediction signal for
the bsize. The tx selection is called in model_rd,
same as for inter-modes.

Adjust threshold in svc datarate test.

Negligible/neutral change in psnr/ssim metrics
for speed 7 and 8, 1 layer and SVC mode.

Change-Id: I33bc8447afdc3785482e13aac5c3636e13c59644