Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > This CL is related to the commit : https://chromium-review.googlesource.com/c/webm/libvpx/+/337710
> > The above commit may put two GF only groups just before a key frame. For example, if the KF is 22 
> > frames away (active_max_gf_interval=16, min_gf_interval=12) then active_max_gf_interval is set to 11 
> > and there is a possibility that the group size is 11. As the group size is lesser than min_gf_interval, 
> > source_alt_ref_pending is set to zero and ARF is not coded. Thus we end up with two GF only groups 
> > just before a key frame.
> > 
> > With the current CL, we make sure that we do not end up with two GF only groups just before a key frame,
> > i.e., we do not set active_max_gf_interval less than min_gf_interval. In the above example, as frames_to_key
> > is less than 2 times min_gf_interval we do not set active_max_gf_interval to 11 thus avoiding two GF only groups.
> > 
> > Please find the BDRATE impact for cpu-used=1 below : (Negative value signifies improvement)
> > 
> > BDRATE (based on GLBPsnr) on an average is -0.40% ranging between 3.60% and -4.36%. 
> > BDRATE (based on AVGPsnr) on an average is -0.34% ranging between 3.76% and -3.72%.
> > BDRATE (based on VPXSSIM) on an average is -0.36% ranging between 3.45% and -3.99%.
> > 
> > Also, encode time reduction is nil for touchdown_pass content averaged at 3, 5, 8 and 10 Mbps.
> 
> What about the case where for example the max_interval is 16 and we are 17 frames away from a key frame. It looks to me like in this case the above code might end up coding an ARF group of 16 and then a single frame GF group.
> 
> I can see the benefit of not setting max < min and so forcing two GF groups but perhaps in this case we should set max = min for the first group and then whatever left for the terminal GF only group.  (So in my example above if min is 12 and max 16 we set max to 12 and then split as 12 , 5 not 16, 1. In you example of a 22 frame group we would then split 12, 10 not anything from 12-16 and then as few as 6 in the final group.

Our understanding is as follows : In ARF group of 12 followed by a GF only group of 5, the second group (of size 5) would be of bad quality as those 5 frames do not use compound prediction. However with GF groups of size 16 and 1, only 1 frame does not use compound prediction. Hence we felt that 16 and 1 would give better quality.

Could you please explain why coding an ARF group of 12 and then a GF group of 5 is better than an ARF group of 16 and a single frame GF group. Is it that bad quality over fewer frames would cause pulsing than lower quality over more number of frames?  

However, we have made the changes as you have suggested and submitted Patchset 4.  
Please find the BD-Rate with Patchset 4. 
BDRATE (based on GLBPsnr) on an average is -0.34% ranging between 3.54% and -4.06%. 
BDRATE (based on AVGPsnr) on an average is -0.31% ranging between 3.69% and -3.88%.
BDRATE (based on VPXSSIM) on an average is -0.36% ranging between 3.33% and -3.49%.

Please merge this patch in case of no issues.

Patch-set: 4
Reviewer: Gerrit User 1273230 <1273230@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>