Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

With this change we are testing secondary reference usage of the frame which is a scene change. It is not necessary that scene-cut frame will have a lower secondary reference usage. But the frame next to scene cut frame will mostly have a low secondary reference usage. Hence we had submitted the original patch.

Is it possible to share the regression issue details(clip, test recipe).

Patch-set: 1
CC: Gerrit User 1177966 <1177966@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1289407 <1289407@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC-email: vikas.prasad@ittiam.com