Start with q=active_best_quality for key frame
We observed that the key frames are mostly encoded at
active best quality or slightly higher than active best
quality. In some cases, we have seen that when the recode
loop start with q higher than active best quality, the
frame is encoded at the borderline of undershooting.
Therefore, for key frames active_best_quality is chosen
as the starting q.

Change-Id: I435d247ab4d1d160f12f5a3710e6cafb5cfd6610