Update patch set 5
Patch Set 5: Code-Review-2

(1 comment)

This appears to rely on using show_existing_frame.  We are aware of this as potentially beneficial in some cases (including multi-layer ARFs), but there are a number of hardware decoders still in use (for example in some TVs) that do not properly support this feature and in some cases crash. Clearly this hardware is not properly bitstream compliant but the fact that it is out there "in the wild" means that unless or until most fail cases are fixed (perhaps via firmware) we cannot use this feature as all of a sudden these devices would fail playing new YouTube content.

Patch-set: 5
Reviewer: Gerrit User 1000899 <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2