Update patch set 2
Patch Set 2:

It seems setting the default cpu-used to anything but 0 (in this change or in vp9_cx_iface.c as the default) causes a crash due to a NULL cpi->Source. My feeling is we should be checking for > 8-bit encodes in set_good_speed_feature_framesize_dependent() in a different way or there's some logic as we transition from 0 to e.g., 4 in the first past vs. starting at 4 that should be generalized.

Patch-set: 2
CC: Gerrit User 1001914 <1001914@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1002219 <1002219@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>