configure: Check if diff exists instead of checking --version.
In some versions of diff (i.e. non-GNU diff), `diff --version` will
exit 1 since they don't recognize `--version`. This ensures that if the
diff command exists so the configure won't error out if the diff version
is not from GNU diffutils.

command(1) is a POSIX utility, so it should exist in a typical UNIX
system.

Change-Id: I73124ef87ff5b1a50886073b7f7fbe0e829dbefd