vp9-svc: Fix to constraint on inter-layer prediction
For fixed/non-flexible SVC mode: condition to constrain
inter-layer prediction (to be from previous spatial layer
of same superframe) was checked for references that
are scaled. Remove the scale check, to be more general.
Include check (for last or if golden is used as long term
reference) that reference is from same spatial layer.

Change-Id: I17bd55cc5f555228d9c023ec6fb11cb47414957b