Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> (1 comment)
> 
> Add the patch but still has overflow issue.

When comparing the C code to sse2 code, the original code is calculating the values correctly and match the C code. The only issue is when the 'mean' is negative then the code would issue a sanitizer warning message. This is because the mean is extracted as 16 bits from mm_registers but it treated as 32 bit positive value which result in Clang warning. This is because the  high order 16 bits are zero in sse2 but in C code they are all FF's. which is a negative value.  If we don't need the accuracy that the C code has, then casting  the "mean" as  short would get correct sign value which is ok too?

Patch-set: 5