bool reader: only add to valid input pointers
Clears a UBSAN warning about incrementing a null buffer by 0.

Preemptively do the same null buffer check for vp9.

Also preemptively move the validation above the assignment as was done
for vp9 in d7efe068d274b6f7b8e743987b6ec84218d9f501

The only way for the this to be reachable is if both source/buffer and
source_sz/size are 0 which is a valid condition.

BUG=chromium:1026961

Change-Id: Iaa480534b49efe1ecc66484b316f8d654e8a1245