Update patch set 6
Patch Set 6:

I don't really know what's valid. It seems like the loop is just extraneous for files with a single partition and this is the correct fix. It's a no-op if it gets called with null buffers. vp8dx_bool_decoder_fill will recognize that the buffer is empty and continue on it's merry way.

Causing an early failure makes this run a lot faster but I'm not sure that's safe. I tested some of the test vectors while developing but there could be a hole in the coverage that I'm just missing. This seems like the safest "fix"

Patch-set: 6