Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> do we need the new encode loop? I think the behavior changed to include the encoder with [1]. previously the expectation was the decode would fail.
> 
> [1] https://chromium-review.googlesource.com/c/webm/libvpx/+/1148806

Yeah so with the change you mentioned, the encoder will fail. I want to catch the error code from the encoder. With new behavior, RunLoop doesn't need decoding in it. Also the old encode function assumes VPX_ENCODE_OK from the encoder, which is not the case for this test now.

Patch-set: 7