Handle the case where arf is not used for a gop.
(1). Add a variable |allow_alt_ref| in cpi, so that we can control
whether to allow alt ref in SimpleEncode.

(2). Change |allow_alt_ref| to be controllable, rather than always 1.

(3). Add a test such that we can use external arf indexes to
reproduce the result of vp9, using gop size info and qps.

Change-Id: Ic03c14d88ed2659df2f57fc9fa6802541dae71e9