Update patch set 1
Patch Set 1:

Just to make sure I understand: currently encoded resolution is in cm->width and cm->height, and then it can be downscaled down to half of that, if we allow downscaling.

Assuming all that, then this works. But there was some code about downscaling to 3/4 of the current resolution. Then current code is too strict and e.g. from 480x360 it will prohibit downscale, while it's still possible to go down with a scale of 3/4.

Patch-set: 1
Reviewer: Gerrit User 1234969 <1234969@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>