Update patch set 1
Patch Set 1:

I think this could be made a lot cleaner by combining the
global sym() PRIVATE
sym():

into 1 define, but there is one place (android compat) where we do sym(rand). Since that *should* only support elf it should be OK to replace that sym() because the only reason we need it is to account for file types that use an _ with the function name and I don't *think* android does that.

Given our switch to intrinsics, I'm not sure how much time it's worth investing here and making sure a large refactor doesn't break something.

Patch-set: 1