Update patch set 3
Patch Set 3:

Thinking about this more last night, I think this is going to be fragile and result in breakages that are hard for folks to understand. E.g., one space gets you a obscure definition error and it's unnatural to drop the space between macro calls.

I really do suggest that instead you adopt the cglobal approach I suggested instead if possible; especially if this is going to become the common fix for both libvpx and libaom.

Patch-set: 3