Update patch set 1
Patch Set 1:

Hey guys, is there any update on this change?

The current encoder behaviour causes troubles when configured to a bitrate of 0 bits per second. From a user perspective it returns success in vp8e_encode() but delivers the previously encoded data which breaks the bitstream. There is no simple work around to detect that.

Using a bitrate of 0 bits per second shouldn't be prohibited and we use it in case of sudden bandwidth drops to reduce accumulated delays.

This change simply clears the output buffer at the start of vp8e_encode() which is probably a good idea to do anyhow 😉

Patch-set: 1
Reviewer: Gerrit User 1402267 <1402267@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>