Refine MMI & MSA detection for mips
Firstly, we added compile check to probe the native
ability of toolchain to decide whether a feature can
to be enabled. Secondly, we added runtime check to
probe features that cpu supports. If MSA and MMI are
both supported, we prefer to MSA. We can configure
and build as following commands:
./configure --cpu=loongson3a && make -j4

Change-Id: I057553216dbc79cfaba9c691d5f4cdab144e1123