vp8: replace gf_noboost_onepass_cbr with CBR condition
The flag gf_noboost_onepass_cbr was used to prevent the
switching of the rate correction factor when the GF is
refreshed without boost, which is the case for CBR mode
(i.e, cpi->oxcf.end_usage == USAGE_STREAM_FROM_SERVER)
But this did not take into account the case where ALT_REF
is refreshed. To make sure the rate correction is not switched
on refresh of GF or ALT_REF for CBR mode, use the CBR condition
explicitly.

Change-Id: I969153fa1ae32c54fa0a724b2c4f2b2621aeb602