Change  zm_factor for Vizier.
Changes the exposed zm_factor parameter.

This patch alters the meaning of the zm_factor
parameter that will be exposed for the Vizier project.

The previous power factor was hard to interpret in terms
of its meaning and effect and has been replaced by a linear factor.
Given that the initial Vizier results suggested a lower zero motion
effect for all formats, the default impact has been reduced.

The patch as it stands gives a modest improvement for PSNR
but is slightly down on some sets for SSIM

(overall psnr, ssim % bdrate change: -ve is better)

lowres       -0.111, 0.001
ugc360p   -0.282, -0.068
midres2    -0.183, 0.059
hdres2       -0.042, 0.172

Change-Id: Ife88fa09215eb4fccea9cbfc2a8abfb281efb28a